+++ src/main/java/bc1/gream/domain/order/service/BuyService.java
@@ -14,12 +14,12 @@ import bc1.gream.domain.order.dto.response.BuyNowResponseDto;
 import bc1.gream.domain.order.entity.Buy;
 import bc1.gream.domain.order.entity.Gifticon;
 import bc1.gream.domain.order.entity.Order;
-import bc1.gream.domain.order.entity.Sell;
+import bc1.gream.domain.sell.entity.Sell;
 import bc1.gream.domain.order.mapper.BuyMapper;
 import bc1.gream.domain.order.mapper.OrderMapper;
 import bc1.gream.domain.order.repository.BuyRepository;
 import bc1.gream.domain.order.repository.OrderRepository;
-import bc1.gream.domain.order.repository.SellRepository;
+import bc1.gream.domain.sell.repository.SellRepository;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.service.query.ProductService;
 import bc1.gream.domain.user.entity.Coupon;
