+++ src/main/java/bc1/gream/domain/sell/service/SellBidService.java
@@ -1,15 +1,15 @@
-package bc1.gream.domain.order.service;
+package bc1.gream.domain.sell.service;
 
 import bc1.gream.domain.gifticon.service.GifticonService;
-import bc1.gream.domain.order.dto.request.SellBidRequestDto;
-import bc1.gream.domain.order.dto.response.SellBidResponseDto;
-import bc1.gream.domain.order.dto.response.SellCancelBidResponseDto;
+import bc1.gream.domain.sell.dto.request.SellBidRequestDto;
+import bc1.gream.domain.sell.dto.response.SellBidResponseDto;
+import bc1.gream.domain.sell.dto.response.SellCancelBidResponseDto;
 import bc1.gream.domain.order.entity.Gifticon;
-import bc1.gream.domain.order.entity.Sell;
-import bc1.gream.domain.order.mapper.SellMapper;
-import bc1.gream.domain.order.repository.SellRepository;
-import bc1.gream.domain.order.service.helper.deadline.Deadline;
-import bc1.gream.domain.order.service.helper.deadline.DeadlineCalculator;
+import bc1.gream.domain.sell.entity.Sell;
+import bc1.gream.domain.sell.mapper.SellMapper;
+import bc1.gream.domain.sell.repository.SellRepository;
+import bc1.gream.domain.sell.service.helper.deadline.Deadline;
+import bc1.gream.domain.sell.service.helper.deadline.DeadlineCalculator;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.user.entity.User;
 import java.time.LocalDate;
