+++ src/test/java/bc1/gream/domain/order/service/SellBidServiceTest.java
@@ -10,10 +10,12 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import bc1.gream.domain.gifticon.service.GifticonService;
-import bc1.gream.domain.order.dto.request.SellBidRequestDto;
-import bc1.gream.domain.order.dto.response.SellBidResponseDto;
-import bc1.gream.domain.order.entity.Sell;
-import bc1.gream.domain.order.repository.SellRepository;
+import bc1.gream.domain.sell.dto.request.SellBidRequestDto;
+import bc1.gream.domain.sell.dto.response.SellBidResponseDto;
+import bc1.gream.domain.sell.entity.Sell;
+import bc1.gream.domain.sell.repository.SellRepository;
+import bc1.gream.domain.sell.service.SellBidService;
+import bc1.gream.domain.sell.service.SellService;
 import bc1.gream.domain.user.entity.User;
 import bc1.gream.test.GifticonTest;
 import java.io.IOException;
